introduction
Energy is strength that is required to sustain physical or mental activities. Energy is playing a
major role in industrial operation. Energy monitoring system used to determine and monitor
active energy value. The purpose of the report is to provide a pepper energy monitoring system
that is suitable for existing architecture and procedures.
If the large scale of corporation has not any computerized tool, the firm could not analyse the
value manually. If the factory is large, it should be more tedious, prone to occur mistakes and
hard to analyse the values. This is a system that overcomes these issues and provides the best
analytical techniques.
The outcomes of the project show the energy values at the point of required time and cost for the
energy that consumed by the either machine or section or plant or whole system. Another objective
is that the system should be suitable for existing architecture and procedures.
The working principle of the project is this system has two major parts. Hardware part and software
part. The hardware part consists of analysers and gateways and communication protocols. Electric
analysers measure the power and energy associated with the machine or department. Also, it
transfers values to gateway through Modbus protocols. Then, gateway converts Modbus data to
MQTT data and uploads it to database. In the software part, there is a web-based application. There
are m users in that application (Admin user, engineering user and viewer). Those users log in to
the web application and the real time energy values, graphs and analyser management as well as
change their user details. This is the working principle of the system.
This report provides deeply information about the project, functions, methodology of developing
and implementation of electric energy management system. Furthermore, compressive analyses
about how do affect this project affects sustainability, cost effectiveness and social impact.